Refactor lookup repositories to remove redundancy.
Prior to this commit, every entity had its own repository for performing lookup calls. The only difference between these repositories was the data type of the entity. This led to redundant code and made the code error prone. The redundancy issue is now resolved by returning the API response as a JSON string to the LiveData.

The observing viewmodels hold another individual live data which has the entity type. Using Transformations from the Lifecycle architecture component, these live data parse the JSON string to the particular entity type. These transformations are triggered when an API response is received. The activities and the fragments observe this instance of the livedata instead of the one in the repository.

There is further scope to reduce redundancy by using generics in ViewModels. The only obstacle is that some entities offer extra functionality. For example, the Artist page show the wikipedia summary if available as well. For releases, the cover art option is available as well.