Improve usage of live data and reactive patterns.
Prior to this commit, the Lookup repository shared a single live data distance across all observers from different activities. This could lead to invalid updates being dispatched to background or inactive activities and fragments. Now, every call to fetchData, fetchWikiSummary and fetchCoverArt returns a new instance of LiveData. This avoids potential bugs.

In accordance with reactive patterns, instead of fetching data from repository and waiting for callbacks the activities observe the live data through viewmodels. When the MBID is set, a call is triggered to fetch data from the API. Using Transformations, the accessory data including cover art and wiki summary is fetched.