Use Dagger Hilt for dependency injection
Most of the ViewModels require a repository to fetch data from the network. Before this commit, the repository were created by calling the static getRepository method inside the ViewModel. While this works fine, it leads to the issue that the ViewModel can no longer be tested in isolation from the repository. This changes with the commit. From now, Hilt will inject the Repository in the ViewModel.