Fix bug due to which some search queries do not load
The diagnosis is that the itemsAfter is required to be at least zero in LoadResult.Page. However, on the last page the value calculated for itemsAfter becomes negative. This causes a silent failure and that page does not load. For entities that have only a single page of entries this is particularly problematic as no page will be loaded. The fix is rather easy. Ensure that itemsAfter is never negative.