Fix bug in LogoutActivity
The LogoutActivity was crashing because of a late call to super.onCreate. super.onCreate should be the first line of the onCreate method or should be called at least before any lifecycle related operations like creating a ViewModel or observing LiveData. The LoginActivity was not registered in the manifest becoming another potential cause of crash. Also, delete the duplicate ic_launcher_foreground.xml.