Temporary fix for WikiData crash
The application always tries to fetch data from wikipedia or wikidata regardless of whether WikiData or Wikipedia link has been found. This can crash the app as the network is unable to handle to empty response and throw a NullPointerException. The issue is actually widespread in the application. Handling error states with LiveData is a whole different ball game which needs to be addressed as a whole for the app.