Introduce LiveData Error Handling
Prior to this, errors were caught locally and logged but not propagated across LiveData. This poses main issues including inconsistent UI states, silent failures and bad UX. This commit introduces the Resource wrapper. This can be used to store the status of the request and wrap the data inside it. The UI can access the status of the data act accordingly.