Do not authorization headers in network requests to wiki sites
The app displays wikipedia for artists and release groups. The network calls to wiki site were also sent with the access token issued by MusicBrainz. This is problematic because media wiki tries to authenticate this token as its own and thus returns an error. Since, the wiki api does not require an access token anyway, the solution is to just skip adding it in the header.