Fetch preference values for Picard IP and port in a getter
Directly assigning the values to the variables makes the properties
behave like static i.e. they are loaded only once during the app's
lifecycle. Instead putting them in a getter ensures that they are
fetched every time, we retrieve the preference. This is useful because
otherwise the user would need to close and reopen the app to see their
come into effect.